Move character counter JS to new component directory
