+++ assets/components/character-counter/character-counter.js
