+++ assets/javascripts/application.js
@@ -44,7 +44,7 @@ var attorneyBanner = require('./modules/attorneyBanner.js')
 var youtubePlayer = require('./modules/youtubePlayer.js')
 var accordion = require('./modules/accordion.js')
 var tabs = require('./modules/tabs.js')
-var charCounter = require('./modules/charCounter.js')
+var charCounter = require('../components/character-counter/character-counter.js')
 var addRemove = require('./modules/addRemove.js')
 var modalDialog = require('./modules/modalDialog.js')
 var dynamicGaTags = require('./modules/dynamicGaTags.js')
